requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi>=0.100.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
pydantic>=2.0.0
sniffio==1.3.0
SQLAlchemy==2.0.10
uvicorn==0.21.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
